Perform refactor. Separate update_graph to different file
